+++ lib/strong_form/tag.rb
@@ -19,12 +19,7 @@ module StrongForm
       return if options.key?(:disabled)
 
       options[:disabled] =
-        object.permitted_attributes != true &&
-        object.permitted_attributes.none? do |entry|
-          entry.is_a?(Hash) &&
-            entry.keys.include?(@method_name.to_sym) ||
-            entry == @method_name.to_sym
-        end
+        !StrongForm::Helper.attribute_permitted?(@method_name, object.permitted_attributes)
     end
   end
 end