+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Always respect explicitly set `disabled` html attribute. (`disabled: false` was ignored)
 * Do correctly enable inputs if permitted attributes includes a hash with multiple keys. (e.g. `permitted_attributes: [:some_attr, { first_key: [], second_key: [] }]`)
+* Add a `#attribute_permitted?` method to a `FormBuilder`, so one can do `f.attribute_permitted?(attr)`
 
 ## 0.0.9